workaround across my mind
-will verify this new thought after work today.
The current code calls the getTwitterTweets() in TwitterComponent, when an item in trends Flatlist is clicked. This causes delay to get the tweets , which does not works perfectly.

Since the trends are obtained through props that passed into the TwitterComponent and works fine, I could've call the getTwitterTweets() for each of the trends, store them into an array then pass it to TwitterComponent.